+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Switched time zone definitions from America/Winnipeg to US/Central
+- Pulled out time zone constants from all over the place into a central constant
 
 ## [2.6.3] - 2018-09-17
 ### Fixed